1. Added ddl_create, ddl_insert and ddl_drop implementations. 2. ddl_insert awaiting fs/heap implementation

+++ src/handler/ddl.h
@@ -1,10 +1,14 @@
 #ifndef DDL_H
 #define DDL_H
 
-#include"db/query.h"
+#include<stdint.h>
 
-int ddl_create(Query *);
-int ddl_insert(Query *);
-int ddl_drop(Query *);
+#include"db/schema.h"
+#include"db/query.h"
+#include"parser/parse_tree.h"
+
+uint8_t ddl_create(Query *);
+uint8_t ddl_insert(Query *);
+uint8_t ddl_drop(Query *);
 
 #endif
